Added dbhelper code to photo repo
