+++ main_project_code/TeamProject/iCollections/Controllers/DatabaseHelper.cs
@@ -142,24 +142,6 @@ namespace iCollections.Controllers
             extractedCollections = extractedCollections.OrderByDescending(r => r.DateMade).ToList();
         }
 
-        public List<PhotoInfo> GetMyPhotosInfo(int myId)
-        {
-            //string address = "https://localhost:5001/api/image/thumbnail/";
-            string address = "https://icollections.azurewebsites.net/api/image/thumbnail/";
-            var photosInformation = _collectionsDbContext.Photos
-                                .Where(row => row.User.Id == myId)
-                                .Select(myRows => new PhotoInfo { Url = address + myRows.PhotoGuid, PhotoName = myRows.Name })
-                                .ToList();
-
-            return photosInformation;
-        }
-
-        public Photo GetPhoto(Guid id)
-        {
-            var photo = _collectionsDbContext.Photos.FirstOrDefault(row => row.PhotoGuid == id);
-            return photo;
-        }
-
         public string GetMyProfilePicUrl(int myId)
         {
             string address = "https://icollections.azurewebsites.net/api/image/thumbnail/";
