+++ main_project_code/TeamProject/iCollections/Data/Abstract/IPhotoRepository.cs
@@ -1,9 +1,13 @@
 using iCollections.Models;
+using System.Collections.Generic;
+using System;
 
 namespace iCollections.Data.Abstract
 {
     public interface IPhotoRepository : IRepository<Photo>
     {
-        
+        List<PhotoInfo> GetMyPhotosInfo(int myId);
+
+        Photo GetPhoto(Guid id);
     }
 }
