+++ main_project_code/TeamProject/iCollections/Data/Concrete/PhotoRepository.cs
@@ -1,5 +1,9 @@
 using iCollections.Models;
 using iCollections.Data.Abstract;
+using System.Collections.Generic;
+using System.Linq;
+using System;
+
 
 namespace iCollections.Data.Concrete
 {
@@ -7,7 +11,24 @@ namespace iCollections.Data.Concrete
     {
         public PhotoRepository(ICollectionsDbContext ctx) : base(ctx)
         {
-            
+
+        }
+
+        public List<PhotoInfo> GetMyPhotosInfo(int myId)
+        {
+            string address = "/api/image/thumbnail/";
+            var photosInformation = GetAll()
+                                .Where(row => row.User.Id == myId)
+                                .Select(myRows => new PhotoInfo { Url = address + myRows.PhotoGuid, PhotoName = myRows.Name })
+                                .ToList();
+
+            return photosInformation;
+        }
+
+        public Photo GetPhoto(Guid id)
+        {
+            var photo = GetAll().FirstOrDefault(row => row.PhotoGuid == id);
+            return photo;
         }
     }
 }
